requirements.txt
fastapi>=0.111.0
anthropic>=0.28.0
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.7.0

# Hybrid retrieval (BM25-only on Vercel; dense disabled via VERCEL=1)
rank-bm25>=0.2.2

# Web content extraction
httpx>=0.27.0
trafilatura>=1.9.0
beautifulsoup4>=4.12.0

# PDF extraction
pypdf>=4.2.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.20.0
alembic>=1.13.0
